Refactor checkout process into separate components
This commit refactors the checkout process into separate components for better code organization, reusability, and maintainability. The following changes have been made:

1. Created a new `BasketManager` component to handle basket creation and package addition logic.
2. Created a new `PackageManager` component to handle package-related operations, such as checking if a specific package is present in the basket.
3. Created a new `TebexCheckout` component to encapsulate the logic related to the Tebex checkout process, including initialization, event handling, and launching the checkout.
4. Refactored the `CheckoutProcess` component to act as the main container for the checkout process flow, orchestrating the interactions between the other components.
5. Updated the import statements and component usage throughout the codebase to reflect the new component structure.

This refactoring improves code organization by separating concerns into individual components, making the codebase more modular and easier to maintain. Each component has a well-defined responsibility, facilitating code reuse and testability.

While the overall functionality remains the same, the code is now more readable, maintainable, and extensible, making it easier to enhance or refactor specific parts of the checkout process without affecting the entire codebase.
