+++ app/components/molecules/CheckoutProcess/BasketManager.tsx
@@ -1,37 +1,36 @@
 // components/molecules/CheckoutProcess/BasketManager.tsx
 
 //In this implementation, the BasketManager component handles the creation of a basket if it doesn't exist. It uses the useFetcherWithPromise hook to submit a POST request to the /store/create route and updates the basketId state with the response.
-import React, { useCallback } from 'react'
-import { useFetcherWithPromise } from '~/utils/general'
+import { useFetcherWithPromise } from '~/utils/general';
 
-interface BasketManagerProps {
-  basketId: string | null
-  setBasketId: (basketId: string | null) => void
-}
+export const createBasket = async () => {
+  const { submit } = useFetcherWithPromise();
 
-const BasketManager: React.FC<BasketManagerProps> = ({ basketId, setBasketId }) => {
-  const { submit } = useFetcherWithPromise()
+  console.log('Creating basket...');
+  try {
+    const response = await submit(null, {
+      method: 'post',
+      action: '/store/create',
+    });
+    return response.basketId;
+  } catch (error) {
+    console.error('Failed to create basket:', error);
+    throw error;
+  }
+};
 
-  const createBasket = useCallback(async () => {
-    console.log('Creating basket...')
-    try {
-      const response = await submit(null, {
-        method: 'post',
-        action: '/store/create',
-      })
-      setBasketId(response.basketId)
-    } catch (error) {
-      console.error('Failed to create basket:', error)
-    }
-  }, [submit, setBasketId])
+export const addPackageToBasket = async (basketId: string) => {
+  const { submit } = useFetcherWithPromise();
 
-  React.useEffect(() => {
-    if (!basketId) {
-      createBasket()
-    }
-  }, [basketId, createBasket])
-
-  return null
-}
-
-export default BasketManager
+  console.log('Adding package to basket...');
+  try {
+    const response = await submit(
+      { basketId },
+      { method: 'post', action: '/store/add' },
+    );
+    return response.packages;
+  } catch (error) {
+    console.error('Failed to add package:', error);
+    throw error;
+  }
+};
